feat: created routes and validation by sanctum

+++ backend/app/Http/Controllers/MembrosController.php
@@ -11,12 +11,18 @@ use Illuminate\Support\Facades\Hash;
 
 class MembrosController extends Controller
 {
-    public function getAll(){
+    public function getAll(Request $request){
+        if ($request->user()->currentAccessToken()['name'] != 'admin_token') {
+            return response(['message' => 'Operação Não Autorizada'], 401);
+        }
         $data = Membros::all();
         return response()->json($data);
     }
 
     public function criarMembros(Request $request){
+        if ($request->user()->currentAccessToken()['name'] != 'admin_token') {
+            return response(['message' => 'Operação Não Autorizada'], 401);
+        }
         $membro = Membros::create($request->validate([
             "nome" => "required",
             "cpf" => "required",
@@ -33,7 +39,7 @@ class MembrosController extends Controller
     public function registrarUsuario(Request $request){
         $membro = DB::select('SELECT id as id_membro, nome, cpf, telefone FROM membros WHERE cpf = "'. $request->validate(['cpf'=>'required'])['cpf'] . '"');
 
-        $m = (array) $membro[0];
+        $m = (array) $membro;
         if(count($membro) == 0){
             return response(['message'=>'Membro não Encontrado'], 404);
         }
@@ -48,17 +54,58 @@ class MembrosController extends Controller
             'senha' => 'required'
         ]);
 
-        $usuario['senha_hash'] = Hash::make($usuario['senha']);
-        $usuario['tipo'] = 'membro';
+        $userData = [
+            'senha_hash' => Hash::make($usuario['senha']),
+            'tipo' => 'membro',
+            'email' => $usuario['email'],
+            'id_membro' => $membro[0]->id_membro,
+            'cpf' => $membro[0]->cpf,
+            'nome' => $membro[0]->nome,
+            'telefone' => $membro[0]->telefone
+        ];
 
-
-        $user = User::create(array_merge($usuario, $m));
-        
+        $user = User::create($userData);
+        // return $userData;
         return response(['data'=> $user], 201);
+    }
+
+    public function registrarAdmin(Request $request){
+        if ($request->user()->currentAccessToken()['name'] != 'admin_token') {
+            return response(['message' => 'Operação Não Autorizada'], 401);
+        }
+
+        $user = $request->validate([
+            'nome'=> 'required',
+            'cpf' => 'required',
+            'telefone' => 'required',
+            'email' => 'required|email',
+            'senha' => 'required'
+        ]);
+
+        $userExists = User::where('cpf', $user['cpf'])->first();
+        
+        if($userExists){
+            return response(['message' => 'Usuario já existe'], 403);
+        }
+
+        $emailValidate = User::where('email', $user['email'])->first();
+        if($emailValidate){
+            return response(['message' => 'Usuario já cadastrado com este email'], 403);
+        }
+
+        $user['tipo'] = 'Admin';
+        $user['senha_hash'] = Hash::make($user['senha']);
+
+        $created = User::create($user);
+
+        return response(['data'=> $created], 201);
     }
 
     public function editMembro(Request $request, $id){
         $membro = Membros::findOrFail($id);
+        if ($request->user()->currentAccessToken()['name'] != 'admin_token' || $request->user()->currentAccessToken()['tokenable']->id != $membro['id']) {
+            return response(['message' => 'Operação Não Autorizada'], 401);
+        }
         $membro->update($request->validate([
             "nome" => "required",
             "endereco" => "required",
@@ -70,7 +117,10 @@ class MembrosController extends Controller
         return response(['data' => $membro], 200);
     }
 
-    public function deleteMembro($id){
+    public function deleteMembro(Request $request,$id){
+        if ($request->user()->currentAccessToken()['name'] != 'admin_token') {
+            return response(['message' => 'Operação Não Autorizada'], 401);
+        }
         $usuario = User::where('id_membro', $id)->first();
         $usuario->delete();
         
@@ -80,7 +130,8 @@ class MembrosController extends Controller
         return response('', 204);
     }
 
-    public function getMe($id){
+    public function getMe(Request $request){
+        $id = $request->user()->currentAccessToken()['tokenable']->id_membro;
         $perfil = Membros::findOrFail($id);
 
         return response(['data'=> $perfil],200);
